requirements.txt
streamlit
google-generativeai
langchain>=0.2.10
langchain-core>=0.2.20
langchain-community>=0.2.10
langchain-google-genai>=0.1.5
faiss-cpu
pypdf
